% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRT2.R
\docType{data}
\name{CRT2}
\alias{CRT2}
\title{Example Data For CRT With 2 Arms}
\format{An object of class \code{data.frame} with 8 rows and 4 columns.}
\usage{
CRT2
}
\description{
\itemize{
 \item ID. The identification number of the subjects.
 \item cluster. The cluster number.
 \item score. The score of the subject.
 \item group. The group number.
}
}
\examples{
#  ID cluster score group
#  1 1 6 0
#  2 1 2 0
#  3 2 6 1
#  4 2 5 1
#  5 3 1 0
#  6 3 4 0
#  7 4 6 1
#  8 4 4 1
}
\keyword{datasets}
