% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRT3.R
\docType{data}
\name{CRT3}
\alias{CRT3}
\title{Example Data For CRT With 3 Arms}
\format{An object of class \code{data.frame} with 30 rows and 4 columns.}
\usage{
CRT3
}
\description{
\itemize{
 \item ID. The identification number of the subjects.
 \item cluster. The cluster number.
 \item score. The score of the subject.
 \item group. The group number.
}
}
\examples{
#  id cluster score group
#  1 1 1.93 0
#  2 1 1.51 0
#  3 1 2.13 0
#  4 1 2.96 0
#  5 1 3.84 0
#  6 2 3.36 1
#  7 2 3.13 1
#  8 2 1.71 1
#  9 2 3.1 1
#  10 2 2.53 1
#  11 3 2.01 2
#  12 3 4.73 2
#  13 3 3.34 2
#  14 3 0.11 2
#  15 3 3.6 2
#  16 4 2 0
#  17 4 1.99 0
#  18 4 1.89 0
#  19 4 2.25 0
#  20 4 1.83 0
#  21 5 3.03 1
#  22 5 2.08 1
#  23 5 1.5 1
#  24 5 3.18 1
#  25 5 1.92 1
#  26 6 3.49 2
#  27 6 3.08 2
#  28 6 4.54 2
#  29 6 2.34 2
#  30 6 4.33 2
}
\keyword{datasets}
