\name{estCRT2arm}
\alias{estCRT2arm}
\alias{estCRT3arm}
\alias{estMRT2arm}
\alias{estMRT3arm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Estimate multilevel effect size from data
}
\description{
Estimate multilevel effect size from data
}
\usage{
estCRT2arm(file)
estCRT3arm(file)
estMRT2arm(file)
estMRT3arm(file)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{file}{
a data file
}
}
