% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model58.R
\name{model58}
\alias{model58}
\title{model58}
\usage{
model58(
  pop.cov = NULL,
  mu = NULL,
  w_value = 0,
  method = "value",
  a1 = 0.5,
  b1 = 0.75,
  c1 = 0.1,
  c2 = 0.1,
  b2 = 0.6,
  b3 = 0.5,
  a2 = 0.4,
  sigx2 = 1,
  sigw2 = 1,
  sige12 = 1,
  sige22 = 1,
  sigx_w = 0.4,
  n = 500,
  nrep = 1000,
  alpha = 0.05,
  b = 1000,
  nb = 300,
  ncore = 5
)
}
\arguments{
\item{pop.cov}{covariance matrix, default to NULL if using the regression coefficient approach}

\item{mu}{mean vector, default to NULL if using the regression coefficient approach}

\item{w_value}{moderator level}

\item{method}{"value" for using the indirect effect value in power calculation, or "joint" for using joint significance in power calculation}

\item{a1}{regression coefficient of outcome (m) on moderator (w)}

\item{b1}{regression coefficient of mediator (m) on predictor (x)}

\item{c1}{regression coefficient of outcome (m) on the product (xw)}

\item{c2}{regression coefficient of outcome (y) on the product (mw)}

\item{b2}{regression coefficient of outcome (y) on mediator (m)}

\item{b3}{regression coefficient of outcome (y) on predictor (x)}

\item{a2}{regression coefficient of outcome (y) on moderator (w)}

\item{sigx2}{variance of predictor (x)}

\item{sigw2}{variance of moderator (w)}

\item{sige12}{variance of error in the first regression equation}

\item{sige22}{variance of error in the second regression equation}

\item{sigx_w}{covariance between predictor (x) and moderator (w)}

\item{n}{sample size}

\item{nrep}{number of replications}

\item{alpha}{type 1 error rate}

\item{b}{number of bootstrap iterations}

\item{nb}{bootstrap sample size}

\item{ncore}{number of cores to use in parallel}
}
\value{
power of indirect effect, direct effect, and moderation
}
\description{
power analysis of model 58 in Introduction to Mediation, Moderation, and Conditional Process Analysis
}
