% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRT2.R
\docType{data}
\name{MRT2}
\alias{MRT2}
\title{Example Data For MRT With 2 Arms}
\format{An object of class \code{data.frame} with 16 rows and 4 columns.}
\usage{
MRT2
}
\description{
\itemize{
 \item ID. The identification number of the subjects.
 \item cluster. The cluster number.
 \item score. The score of the subject.
 \item group. The group number.
}
}
\examples{
#Example data for MRT with 2 arms
#  id cluster score group
#  1 1 6 0
#  2 1 2 0
#  3 1 3 1
#  4 1 3 1
#  5 2 6 0
#  6 2 10 0
#  7 2 7 1
#  8 2 6 1
#  9 3 6 0
#  10 3 5 0
#  11 3 4 1
#  12 3 4 1
#  13 4 1 0
#  14 4 8 0
#  15 4 10 1
#  16 4 -2 1
}
\keyword{datasets}
