% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRT3.R
\docType{data}
\name{MRT3}
\alias{MRT3}
\title{Example Data For MRT With 3 Arms}
\format{An object of class \code{data.frame} with 24 rows and 4 columns.}
\usage{
MRT3
}
\description{
\itemize{
 \item ID. The identification number of the subjects.
 \item cluster. The cluster number.
 \item score. The score of the subject.
 \item group. The group number.
}
}
\examples{
#  id cluster score group
#  1 1 2 0
#  2 1 3 0
#  3 1 2 1
#  4 1 0 1
#  5 1 3 2
#  6 1 2 2
#  7 2 1 0
#  8 2 4 0
#  9 2 2 1
#  10 2 3 1
#  11 2 3 2
#  12 2 1 2
#  13 3 1 0
#  14 3 4 0
#  15 3 1 1
#  16 3 1 1
#  17 3 2 2
#  18 3 0 2
#  19 4 4 0
#  20 4 3 0
#  21 4 1 1
#  22 4 3 1
#  23 4 3 2
#  24 4 3 2
}
\keyword{datasets}
