% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webpower.R
\name{plot.webpower}
\alias{plot.webpower}
\title{To plot Statistical Power Curve}
\usage{
\method{plot}{webpower}(x, xvar = NULL, yvar = NULL, xlab = NULL,
  ylab = NULL, ...)
}
\arguments{
\item{x}{Objects of power analysis.}

\item{xvar}{The variable name used as the x (horizontal) axis. It is not required.}

\item{yvar}{The variable name used as the y (vertical) axis. It is not required.}

\item{xlab}{The label for the x axis. It is not required.}

\item{ylab}{The label for the y axis. It is not required.}

\item{...}{Extra arguments. It is not required.}
}
\value{
The plot.
}
\description{
This function is used to plot the power curves generated by webpower.
}
\examples{
res <- wp.correlation(n=seq(50,100,10),r=0.3, alternative="two.sided")
plot(res)
}
