% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webpower.R
\docType{package}
\name{print.webpower}
\alias{print.webpower}
\title{To Print Statistical Power Analysis Results}
\usage{
\method{print}{webpower}(x, ...)
}
\arguments{
\item{x}{Object of power analysis. It is an object returned by a webpower function such as wp.anova().}

\item{...}{Extra arguments. It is not required.}
}
\value{
The printing of the input object of power analysis.
}
\description{
This function is used to summary the power analysis results.
}
\examples{
res <- wp.correlation(n=50,r=0.3, alternative="two.sided")
print(res)

}
