% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webpower.R
\name{summary.power}
\alias{summary.power}
\title{Summary Statistical Power Analysis Results}
\usage{
\method{summary}{power}(object, ...)
}
\arguments{
\item{object}{Object of power analysis. It is an object returned by a webpower function for SEM based on Monte Carlo methods with class = 'power'.}

\item{...}{Extra arguments. It is not required.}
}
\value{
The summary of the input object of power analysis.
}
\description{
This function is used to summary the power analysis results.
}
