\name{WebPower-package}
\alias{WebPower-package}
\alias{WebPower}
\docType{package}
\title{
Basic and Advanced Statistical Power Analysis
}
\description{
This is a collection of tools for conducting both basic and advanced statistical power analysis including correlation, proportion, t-test, one-way ANOVA, two-way ANOVA, linear regression, logistic regression, Poisson regression, mediation analysis, longitudinal data analysis, structural equation modeling and multilevel modeling. It also serves as the engineer for conducting power analysis online at https://webpower.psychstat.org.
}
\details{
This is a collection of tools for conducting both basic and advanced statistical power analysis including correlation, proportion, t-test, one-way ANOVA, two-way ANOVA, linear regression, logistic regression, Poisson regression, mediation analysis, longitudinal data analysis, structural equation modeling and multilevel modeling. It also serves as the engineer for conducting power analysis online at https://webpower.psychstat.org.
}

\references{
Zhang, Z., & Yuan, K.-H. (2018). Practical Statistical Power Analysis Using Webpower and R (Eds). Granger, IN: ISDSA Press.
}
