% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model7.R
\name{wp.modmed.m7}
\alias{wp.modmed.m7}
\title{model7}
\usage{
wp.modmed.m7(
  beta1 = 0.39,
  beta2 = 0.2,
  beta3 = 0.3,
  a1 = 0.39,
  c1 = 0.2,
  sigx2 = 1,
  sigw2 = 1,
  sige12 = 1,
  sige22 = 1,
  sigx_w = 0.5,
  n = 500,
  nrep = 1000,
  alpha = 0.05,
  b = 1000,
  nb = n,
  w_value = 0,
  method = "value",
  ncore = 1,
  pop.cov = NULL,
  mu = NULL,
  varnames = c("y", "x", "w", "m", "xw")
)
}
\arguments{
\item{beta1}{regression coefficient of mediator (m) on predictor (x)}

\item{beta2}{regression coefficient of outcome (y) on predictor (x)}

\item{beta3}{regression coefficient of outcome (y) on mediator (m)}

\item{a1}{regression coefficient of mediator (m) on moderator (w)}

\item{c1}{regression coefficient of mediator (m) on the product (xw)}

\item{sigx2}{variance of predictor (x)}

\item{sigw2}{variance of moderator (w)}

\item{sige12}{variance of error in the first regression equation}

\item{sige22}{variance of error in the second regression equation}

\item{sigx_w}{covariance between predictor (x) and moderator (w)}

\item{n}{sample size}

\item{nrep}{number of replications}

\item{alpha}{type 1 error rate}

\item{b}{number of bootstrap iterations}

\item{nb}{bootstrap sample size, default to n}

\item{w_value}{moderator level}

\item{method}{"value" for using the indirect effect value in power calculation, or "joint" for using joint significance in power calculation}

\item{ncore}{number of cores to use, default is 1, when ncore > 1, parallel is used}

\item{pop.cov}{covariance matrix, default to NULL if using the regression coefficient approach}

\item{mu}{mean vector, default to NULL if using the regression coefficient approach}

\item{varnames}{name of variables for the covariance matrix}
}
\value{
power of indirect effect, direct effect, and moderation
}
\description{
power analysis of model 7 in Introduction to Mediation, Moderation, and Conditional Process Analysis
}
\examples{
# usage of wp.modmed.m7
test = wp.modmed.m7(beta1 = 0.39, beta2 = 0.2, beta3 = 0.3, a1 = 0.39,
          c1 = 0.2, sigx2 = 1, sigw2 = 1, sige12 = 1,
         sige22 = 1, sigx_w = 0.5, n = 500, nrep = 200,
         alpha = 0.05, b = 1000, ncore = 1)
print(test)
}
