% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webpower.R
\name{wp.popPar}
\alias{wp.popPar}
\title{Extract Population Value Table}
\usage{
wp.popPar(object)
}
\arguments{
\item{object}{Object of power analysis. It is an object returned by a webpower function for SEM based on Monte Carlo methods with class = 'power'.}
}
\value{
Population value table of parameters from the input object of power analysis.
}
\description{
This function is used to extract population value table for parameters form an power analysis object for SEM based on Monte Carlo methods (class = 'power').
}
