% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webpower.R
\name{plot.lcs.power}
\alias{plot.lcs.power}
\title{Plot the power curve for Latent Change Score Models}
\usage{
\method{plot}{lcs.power}(x, parameter, ...)
}
\arguments{
\item{x}{Data to plot.}

\item{parameter}{Parameters for features of the plot.}

\item{...}{Extra arguments. It is not required.}
}
\description{
This function is used to plot the power analysis results for Latent Change Score Models.
}
\references{
Zhang, Z., & Liu, H. (2018). Sample Size and Measurement Occasion Planning for Latent Change Score Models through Monte Carlo Simulation. In E. Ferrer, S. M. Boker, and K. J. Grimm (Eds.) Advances in Longitudinal Models for Multivariate Psychology: A Festschrift for Jack McArdle.

Zhang, Z., & Yuan, K.-H. (2018). Practical Statistical Power Analysis Using Webpower and R (Eds). Granger, IN: ISDSA Press.
}