% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model8.R
\name{wp.modmed.m8}
\alias{wp.modmed.m8}
\title{model8}
\usage{
wp.modmed.m8(
  beta1 = 0.2,
  beta2 = 0.75,
  beta3 = 0.8,
  a1 = 0.6,
  a2 = 0.3,
  c1 = 0.8,
  c2 = 0.79,
  sigx2 = 1,
  sigw2 = 1,
  sige12 = 1,
  sige22 = 1,
  sigx_w = 0.5,
  n = 100,
  nrep = 1000,
  alpha = 0.05,
  b = 1000,
  nb = n,
  w_value = 0,
  method = "value",
  ncore = 1,
  pop.cov = NULL,
  mu = NULL,
  varnames = c("y", "x", "w", "m", "xw")
)
}
\arguments{
\item{beta1}{regression coefficient of mediator (m) on predictor (x)}

\item{beta2}{regression coefficient of outcome (y) on predictor (m)}

\item{beta3}{regression coefficient of outcome (y) on mediator (x)}

\item{a1}{regression coefficient of mediator (m) on moderator (w)}

\item{a2}{regression coefficient of mediator (y) on moderator (w)}

\item{c1}{regression coefficient of mediator (m) on the product (xw)}

\item{c2}{regression coefficient of mediator (y) on the product (xw)}

\item{sigx2}{variance of predictor (x)}

\item{sigw2}{variance of moderator (w)}

\item{sige12}{variance of error in the first regression equation}

\item{sige22}{variance of error in the second regression equation}

\item{sigx_w}{covariance between predictor (x) and moderator (w)}

\item{n}{sample size}

\item{nrep}{number of replications}

\item{alpha}{type 1 error rate}

\item{b}{number of bootstrap iterations}

\item{nb}{bootstrap sample size, default to n}

\item{w_value}{moderator level}

\item{method}{"value" for using the indirect effect value in power calculation, or "joint" for using joint significance in power calculation}

\item{ncore}{number of cores to use, default is 1, when ncore > 1, parallel is used}

\item{pop.cov}{covariance matrix, default to NULL if using the regression coefficient approach}

\item{mu}{mean vector, default to NULL if using the regression coefficient approach}

\item{varnames}{name of variables for the covariance matrix}
}
\value{
power of indirect effect, direct effect, and moderation
}
\description{
power analysis of model 8 in Introduction to Mediation, Moderation, and Conditional Process Analysis
}
\examples{
# usage of wp.modmed.m8
test = wp.modmed.m8(beta1 = 0.2, beta2 = 0.2, beta3 = 0.2,
                   a1 = 0.2, a2=0.2, c1 = 0.2, c2 = 0.2,
                   sigx2 = 1, sigw2 = 1, sige12 = 1, sige22 = 1, sigx_w = 0.5,
                   nrep = 50, alpha = 0.05, b = 1000, ncore = 1)
print(test)
}
